inject & mock CommandClient factory and DiscoverProcessWorker, tests mocked up to wifi selection
