+++ devicesetup/src/main/java/io/particle/android/sdk/devicesetup/ui/ConnectToApFragment.java
@@ -13,6 +13,8 @@ import io.particle.android.sdk.devicesetup.ParticleDeviceSetupLibrary;
 import io.particle.android.sdk.di.ApModule;
 import io.particle.android.sdk.utils.EZ;
 import io.particle.android.sdk.utils.SSID;
+import io.particle.android.sdk.utils.SoftAPConfigRemover;
+import io.particle.android.sdk.utils.WifiFacade;
 import io.particle.android.sdk.utils.WorkerFragment;
 import io.particle.android.sdk.utils.ui.Ui;
 
@@ -36,7 +38,9 @@ public class ConnectToApFragment extends WorkerFragment {
         return frag;
     }
 
-    @Inject protected ApConnector apConnector;
+    protected ApConnector apConnector;
+    @Inject protected SoftAPConfigRemover configRemover;
+    @Inject protected WifiFacade wifiFacade;
     private Client apConnectorClient;
 
     @Override
@@ -50,6 +54,7 @@ public class ConnectToApFragment extends WorkerFragment {
         super.onCreate(savedInstanceState);
         ParticleDeviceSetupLibrary.getInstance().getApplicationComponent().activityComponentBuilder()
                 .apModule(new ApModule()).build().inject(this);
+        apConnector = new ApConnector(getActivity(), configRemover, wifiFacade);
     }
 
     @Override
