+++ devicesetup/src/main/java/io/particle/android/sdk/di/ApModule.java
@@ -5,7 +5,8 @@ import android.support.annotation.RestrictTo;
 
 import dagger.Module;
 import dagger.Provides;
-import io.particle.android.sdk.devicesetup.ApConnector;
+import io.particle.android.sdk.devicesetup.commands.CommandClientFactory;
+import io.particle.android.sdk.devicesetup.ui.DiscoverProcessWorker;
 import io.particle.android.sdk.utils.SoftAPConfigRemover;
 import io.particle.android.sdk.utils.WifiFacade;
 
@@ -19,12 +20,17 @@ public class ApModule {
     }
 
     @Provides
-    protected ApConnector providesApConnector(Context context, SoftAPConfigRemover configRemover, WifiFacade wifiFacade) {
-        return new ApConnector(context, configRemover, wifiFacade);
+    protected WifiFacade providesWifiFacade(Context context) {
+        return WifiFacade.get(context);
     }
 
     @Provides
-    protected WifiFacade providesWifiFacade(Context context) {
-        return WifiFacade.get(context);
+    protected DiscoverProcessWorker providesDiscoverProcessWorker() {
+        return new DiscoverProcessWorker();
+    }
+
+    @Provides
+    protected CommandClientFactory providesCommandClientFactory() {
+        return new CommandClientFactory();
     }
 }
