+++ testapp/build.gradle
@@ -1,0 +1,52 @@
+apply plugin: 'com.android.application'
+
+android {
+    compileSdkVersion 26
+    buildToolsVersion "26.0.1"
+
+    defaultConfig {
+        applicationId "io.particle.devicesetup.testapp"
+        minSdkVersion 15
+        targetSdkVersion 26
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.android.support:appcompat-v7:26.0.0'
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+
+    compile project(':devicesetup')
+
+    testCompile "junit:junit:4.12"
+    testCompile 'com.github.fabioCollini:DaggerMock:0.7.0'
+    testCompile "org.mockito:mockito-core:2.8.8"
+    testCompile "com.google.dagger:dagger-compiler:2.11"
+
+    androidTestCompile 'com.android.support:support-annotations:26.0.2'
+    androidTestCompile "org.mockito:mockito-core:2.8.8"
+    androidTestCompile "org.mockito:mockito-android:2.8.8"
+    androidTestCompile "com.android.support.test:runner:0.5"
+    androidTestCompile "com.android.support.test:rules:0.5"
+    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+        exclude group: 'com.google.code.findbugs', module: 'jsr305'
+    })
+    androidTestCompile 'com.github.fabioCollini:DaggerMock:0.7.0'
+}
+
+repositories {
+    maven {
+        url 'https://maven.google.com'
+    }
+    maven { url "https://jitpack.io" }
+}
