+++ testapp/src/main/java/io/particle/devicesetup/testapp/MainActivity.java
@@ -1,0 +1,13 @@
+package io.particle.devicesetup.testapp;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
